+++ front-week06/src/pages/QuizApp.jsx
@@ -1,11 +1,6 @@
-import { useCallback, useRef, useState } from "react"; // ① useRef import 하기
+import { useQuiz } from "../hooks/useQuiz";
 
 const QuizApp = () => {
-  const [currentQuestion, setCurrentQuestion] = useState(0);
-  const [userAnswer, setUserAnswer] = useState("");
-  const [answers, setAnswers] = useState([]);
-  const inputRef = useRef();
-
   const questions = [
     { question: "React에서 상태를 관리하는 Hook은?", answer: "useState" },
     {
@@ -15,28 +10,15 @@ const QuizApp = () => {
     { question: "DOM에 직접 접근할 때 사용하는 Hook은?", answer: "useRef" },
   ];
 
-  const handleSubmit = useCallback(() => {
-    const newAnswers = [...answers];
-    newAnswers[currentQuestion] = userAnswer;
-    setAnswers(newAnswers);
-
-    // ④-1. 다음 문제로 이동하면서 input에 포커스
-    if (currentQuestion < questions.length - 1) {
-      setCurrentQuestion((prev) => prev + 1);
-      setUserAnswer("");
-      setTimeout(() => inputRef.current?.focus(), 0); // (주석처리 풀면 OK)
-    }
-  }, [currentQuestion, userAnswer, answers, questions.length]);
-
-  const handleReset = useCallback(() => {
-    setCurrentQuestion(0);
-    setUserAnswer("");
-    setAnswers([]);
-    inputRef.current.value = "";
-    inputRef.current.focus();
-
-    // ④-3. input에 포커스 하기
-  }, []);
+  const {
+    currentQuestion,
+    userAnswer,
+    setUserAnswer,
+    answers,
+    inputRef,
+    handleSubmit,
+    handleReset,
+  } = useQuiz(questions);
 
   return (
     <div className="quiz-container">
